pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discharge_docs"
version = "2.9.0"
authors = [
    { name = "Ruben Peters", email = "r.peters-7@umcutrecht.nl" },
    { name = "Laura Veerhoek", email = "l.p.veerhoek@umcutrecht.nl" },
    { name = "Welmoed Tjepkema", email= "w.w.a.tjepkema@umcutrecht.nl"}
]
description = "Tool that uses LLM's to automatically generate discharge documentation. Part of the 'AI for administration load' project."
readme = "README.md"
requires-python = ">= 3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.8",
    "python-dotenv>=1.0",
    "deduce>=3.0.0",
    "matplotlib>=3.8",
    "plotly>=5.18",
    "dash-bootstrap-components>=2.0",
    "pandas>=2.1",
    "SQLAlchemy>=2.0",
    "tiktoken>=0.5",
    "pyarrow>=14.0",
    "dash>=3.0",
    "fastapi>=0.111.0",
    "pymssql>=2.3.0",
    "striprtf>=0.0.26",
    "tomli-w>=1.2.0",
    "rich>=13.9.4",
    "altair>=5.5.0",
    "streamlit>=1.44.0",
    "email-validator>=2.2.0",
    "umcu-ai-utils>=0.2.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "nbstripout>=0.7.1",
    "pytest>=8.0.2",
    "uvicorn>=0.30.1",
    "pytest-asyncio>=0.23.6",
    "rsconnect-python>=1.24.0",
    "nltk>=3.9.1",
    "scipy>=1.15.1",
    "pytest-cov>=6.0.0",
    "ipywidgets>=8.1.5",
    "dvc>=3.59.1",
    "alembic>=1.15.2",
]
lint = ["ruff>=0.11.0"]

[tool.isort]
profile = "black"

[tool.ruff]
src = ["src"]

[tool.ruff.lint]
select = ["B", "E", "F", "W", "C90", "I"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "Depends",
    "fastapi.Depends",
    "fastapi.params.Depends",
    "Security",
]

[tool.uv]
default-groups = ["dev", "lint"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
